Refactor QF
