+++ app/managers/context/lib/layers/Context.js
@@ -1,5 +1,5 @@
 
-const Layer = require('../lib/Layer')
+const Layer = require('../Layer')
 
 class Context extends Layer {
