+++ app/managers/context/lib/layers/Workspace.js
@@ -1,5 +1,5 @@
 
-const Layer = require('../lib/Layer')
+const Layer = require('../Layer')
 
 class Workspace extends Layer {
